+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_params.py
@@ -12,7 +12,7 @@ from azure.cli.core.commands.validators import get_default_location_from_resourc
 from azure.mgmt.containerinstance.models import (
     ContainerGroupRestartPolicy, OperatingSystemTypes, ContainerNetworkProtocol)
 from ._validators import (validate_volume_mount_path, validate_secrets, validate_subnet,
-                          validate_gitrepo_directory, validate_network_profile)
+                          validate_gitrepo_directory, validate_network_profile, validate_image)
 
 # pylint: disable=line-too-long
 
@@ -63,7 +63,7 @@ def load_arguments(self, _):
 
     with self.argument_context('container create') as c:
         c.argument('location', arg_type=get_location_type(self.cli_ctx), validator=get_default_location_from_resource_group)
-        c.argument('image', help='The container image name')
+        c.argument('image', validator=validate_image, help='The container image name')
         c.argument('cpu', type=int, help='The required number of CPU cores of the containers, accurate to one decimal place')
         c.argument('memory', type=float, help='The required memory of the containers in GB, accurate to one decimal place')
         c.argument('os_type', arg_type=get_enum_type(OperatingSystemTypes), help='The OS type of the containers')