+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -3,9 +3,12 @@
 Release History
 ===============
 
-0.3.6
+0.3.7
 +++++
 * Add '--assign-identity' for adding a MSI identity to a container group
+
+0.3.6
++++++
 * Show warning when creating a container group with an image without a long running process
 * Fix table output issues for 'list' and 'show' commands